feat: get react-native version from `node_modules`
