+++ packages/cli/src/commands/init/init.ts
@@ -28,10 +28,9 @@ import {getBunVersionIfAvailable} from '../../tools/bun';
 import {getNpmVersionIfAvailable} from '../../tools/npm';
 import {getYarnVersionIfAvailable} from '../../tools/yarn';
 import {createHash} from 'crypto';
-import execa from 'execa';
+import createGitRepository from './createGitRepository';
 
 const DEFAULT_VERSION = 'latest';
-const packageJson = require('./../../../package.json');
 
 type Options = {
   template?: string;
@@ -308,53 +307,6 @@ function userAgentPackageManager() {
 
   return null;
 }
-
-const createGitRepository = async (folder: string) => {
-  const loader = getLoader();
-
-  try {
-    await execa('git', ['--version'], {stdio: 'ignore'});
-  } catch {
-    loader.fail('Unable to initialize Git repo. `git` not in $PATH.');
-    return;
-  }
-
-  try {
-    await execa('git', ['rev-parse', '--is-inside-work-tree'], {
-      stdio: 'ignore',
-      cwd: folder,
-    });
-    loader.succeed(
-      'New project is already inside of a Git repo, skipping git init.',
-    );
-    return;
-  } catch {}
-
-  loader.start('Initializing Git repository');
-
-  try {
-    await execa('git', ['init'], {cwd: folder});
-    await execa('git', ['branch', '-M', 'main'], {cwd: folder});
-    await execa('git', ['add', '.'], {cwd: folder});
-    await execa(
-      'git',
-      [
-        'commit',
-        '-m',
-        `Initial commit\n\nGenerated by ${packageJson.name} ${packageJson.version}.`,
-      ],
-      {
-        cwd: folder,
-      },
-    );
-    loader.succeed();
-  } catch (e) {
-    loader.fail(
-      'Could not create an empty Git repository, see debug logs with --verbose',
-    );
-    logger.debug(e as string);
-  }
-};
 
 export default (async function initialize(
   [projectName]: Array<string>,
